+++ src/main/webapp/package.json
@@ -8,7 +8,7 @@
   },
   "husky": {
     "hooks": {
-      "pre-receive": "lint-staged"
+      "pre-commit": "pretty-quick --staged"
     }
   },
   "lint-staged": {
@@ -122,7 +122,7 @@
     "eslint-plugin-react-hooks": "^2.3.0",
     "expose-loader": "^0.7.5",
     "file-loader": "^5.0.2",
-    "husky": "^3.1.0",
+    "husky": "^4.2.3",
     "lint-staged": "^9.5.0",
     "mini-css-extract-plugin": "^0.8.1",
     "node-sass": "^4.12.0",
