+++ src/main/webapp/resources/js/components/Header/PageHeader.jsx
@@ -3,12 +3,7 @@ import { render } from "react-dom";
 import { Session } from "../session/Session";
 import { Notifications } from "../notifications/Notifications";
 import GalaxyAlert from "./GalaxyAlert";
-import { Breadcrumb } from "antd";
-import { HomeTwoTone } from "@ant-design/icons";
-import { setBaseUrl } from "../../utilities/url-utilities";
-import styled from "styled-components";
-import { blue1 } from "../../styles/colors";
-import { SPACE_XS } from "../../styles/spacing";
+import { Breadcrumbs } from "./Breadcrumbs";
 
 /*
 WEBPACK PUBLIC PATH:
@@ -17,11 +12,6 @@ the variable `__webpack_public_path__`
 See: https://webpack.js.org/guides/public-path/#on-the-fly
  */
 __webpack_public_path__ = `dist/`;
-
-const BreadCrumbs = styled(Breadcrumb)`
-  background-color: ${blue1};
-  padding: ${SPACE_XS};
-`;
 
 export class PageHeader extends React.Component {
   state = {
@@ -37,20 +27,7 @@ export class PageHeader extends React.Component {
   render() {
     return (
       <div>
-        {window.breadcrumbs ? (
-          <BreadCrumbs>
-            <Breadcrumb.Item>
-              <a href={setBaseUrl("")}>
-                <HomeTwoTone />
-              </a>
-            </Breadcrumb.Item>
-            {window.breadcrumbs.map(crumb => (
-              <Breadcrumb.Item key={crumb.label}>
-                <a href={crumb.url}>{crumb.label}</a>
-              </Breadcrumb.Item>
-            ))}
-          </BreadCrumbs>
-        ) : null}
+        <Breadcrumbs crumbs={window.breadcrumbs} />
         <Session />
         <Notifications />
         {this.state.inGalaxy ? <GalaxyAlert /> : null}
